ci: harden CI test pipeline against ephemeral failures (BLDX-1471)
Three-layer approach to prevent flaky network/setup failures from
blocking auto-merge:

Layer A — step-level retry (primary fix for 504/network on download):
- Add .github/scripts/with-retry.sh: externalized, tested bash retry
  script (3 attempts, linear backoff, env-tunable). Backed by a bats
  unit test suite (.github/scripts/tests/with-retry.bats).
- setup-deps: wrap uv sync in with-retry.sh — fixes ephemeral
  failures for every job (unit matrix, integration, storage, connectors)
- sdk-tests-reusable: retried docker pulls for all emulator images
- connector-integration-tests: retry Dapr/Temporal CLI downloads and
  component fetches

Layer B — pytest-rerunfailures on integration/storage/connector jobs:
- Add pytest-rerunfailures>=15.0,<16.0 and pytest-socket>=0.7.0,<1.0.0
  to the test dependency group; regenerate uv.lock
- integration + storage-emulator jobs: --reruns 2 --reruns-delay 5
- connector-integration-tests: pytest-rerunfailures installed transiently
  alongside xdist/cov; default pytest-args updated with --reruns 2
- Unit job: NO reruns (hermetic — failures are real bugs)

Hermeticity enforcement — prevent unit suite from masking real failures:
- unit-tests action: add --disable-socket --allow-unix-socket to enforce
  that unit tests make zero real network calls
- Audit surfaced 25 non-hermetic unit tests:
  - 17 moved to tests/integration/{server,clients,dev,testing}/
    (real local TCP servers, TLS connections, socket port-binding)
  - 8 fixed in place by completing existing mocks
    (TestEmbeddedDaprLifecycle: stub _pick_free_port;
     test_kube_http_call_*: stub _find_free_port)
- Result: 5163 unit tests pass cleanly under --disable-socket
