+++ .github/workflows/sdk-tests-reusable.yaml
@@ -66,7 +66,7 @@ jobs:
             embedded-runtimes-${{ runner.os }}-
 
       - name: Run integration tests
-        run: uv run pytest tests/integration/ -m integration -v --timeout=120
+        run: uv run pytest tests/integration/ -m integration -v --timeout=120 --reruns 2 --reruns-delay 5
 
   storage-emulator:
     name: Storage Emulator Tests
@@ -79,6 +79,11 @@ jobs:
       - name: Start object-store emulators
         run: |
           set -euo pipefail
+          # Pull images first with retry — image pulls are the main 504 surface.
+          .github/scripts/with-retry.sh docker pull minio/minio:RELEASE.2025-09-07T16-13-09Z
+          .github/scripts/with-retry.sh docker pull mcr.microsoft.com/azure-storage/azurite:3.31.0
+          .github/scripts/with-retry.sh docker pull gcr.io/cloud-devrel-public-resources/storage-testbench:v0.62.0
+          .github/scripts/with-retry.sh docker pull nginx:1.27-alpine
           # S3 → MinIO
           docker run -d --name minio -p 9000:9000 \
             -e MINIO_ROOT_USER=minioadmin -e MINIO_ROOT_PASSWORD=minioadmin \
@@ -119,7 +124,7 @@ jobs:
           AWS_ENDPOINT_URL: http://localhost:9000
           AZURE_STORAGE_ENDPOINT: http://localhost:10000
           GCS_EMULATOR_ENDPOINT: http://localhost:9095
-        run: uv run pytest tests/integration/storage -m storage_emulator -v --timeout=300
+        run: uv run pytest tests/integration/storage -m storage_emulator -v --timeout=300 --reruns 2 --reruns-delay 5
 
       - name: Teardown emulators
         if: always()
