+++ pyproject.toml
@@ -145,6 +145,8 @@ test = [
     "coverage>=7.6.1,<8.0.0",
     "pytest-asyncio>=1.4.0,<2.0.0",
     "pytest-timeout>=2.3.0,<3.0.0",
+    "pytest-rerunfailures>=15.0,<16.0",
+    "pytest-socket>=0.7.0,<1.0.0",
     "hypothesis>=6.114.0,<7.0.0",
     "trustme>=1.2.1,<2.0.0",
     "pyjwt[crypto]>=2.12.0",
@@ -181,6 +183,7 @@ stop-deps.shell = "lsof -ti:3000,3500,7233,8233,50001 | xargs kill -9 2>/dev/nul
 # `temporal server start-dev` or daprd needed. Set TEMPORAL_HOST to point at
 # an already-running server instead.
 test-integration = "pytest tests/integration/ -m integration -v --timeout=120"
+test-scripts = "bats .github/scripts/tests"
 
 # API docs — must be before any sub-table header ([tool.poe.tasks.*])
 # or TOML nests it inside the sub-table rather than the parent tasks namespace
