+++ tests/unit/clients/test_ssl_utils.py
@@ -9,7 +9,6 @@ import aiohttp
 import httpx
 import pytest
 import trustme
-from aiohttp import web
 
 from application_sdk.clients.ssl_utils import (
     create_ssl_context_with_custom_certs,
@@ -97,13 +96,15 @@ class TestGetSslContext:
 
     def test_returns_ssl_context_when_cert_dir_set(self):
         """Test that SSLContext is returned when SSL_CERT_DIR is set."""
-        with tempfile.TemporaryDirectory() as tmpdir:
-            with patch(
+        with (
+            tempfile.TemporaryDirectory() as tmpdir,
+            patch(
                 "application_sdk.clients.ssl_utils.get_ssl_cert_dir",
                 return_value=tmpdir,
-            ):
-                result = get_ssl_context()
-                assert isinstance(result, ssl.SSLContext)
+            ),
+        ):
+            result = get_ssl_context()
+            assert isinstance(result, ssl.SSLContext)
 
     def test_integration_with_real_directory(self):
         """Integration test with real temporary directory."""
@@ -291,163 +292,3 @@ class TestGetCustomCaCertBytes:
                 assert b"-----BEGIN CERTIFICATE-----" in result
                 assert b"-----END CERTIFICATE-----" in result
                 assert result.count(b"-----BEGIN CERTIFICATE-----") > 1
-
-
-@pytest.fixture
-def trustme_ca():  # type: ignore[no-untyped-def]
-    """Create a trustme CA for testing."""
-    return trustme.CA()
-
-
-@pytest.fixture
-def server_ssl_ctx(trustme_ca):  # type: ignore[no-untyped-def]
-    """Create a server SSL context with a certificate from the trustme CA."""
-    server_ssl_ctx = ssl.SSLContext(ssl.PROTOCOL_TLS_SERVER)
-    trustme_ca.issue_cert("localhost", "127.0.0.1").configure_cert(server_ssl_ctx)  # type: ignore[no-untyped-call]
-    return server_ssl_ctx
-
-
-class TestSslContextWithPrivateServer:
-    """Integration tests with a real HTTPS server using private certificates."""
-
-    @pytest.mark.asyncio
-    async def test_httpx_connects_to_private_https_server(
-        self, trustme_ca, server_ssl_ctx
-    ):  # type: ignore[no-untyped-def]
-        """Test that httpx can connect to an HTTPS server using a private CA certificate."""
-        with tempfile.TemporaryDirectory() as tmpdir:
-            ca_cert_path = f"{tmpdir}/ca.pem"
-            trustme_ca.cert_pem.write_to_path(ca_cert_path)  # type: ignore[no-untyped-call]
-
-            async def handle_request(request):  # type: ignore[no-untyped-def]
-                return web.Response(text="Hello from private HTTPS server!")
-
-            app = web.Application()
-            app.router.add_get("/", handle_request)
-
-            runner = web.AppRunner(app)
-            await runner.setup()
-
-            site = web.TCPSite(runner, "127.0.0.1", 0, ssl_context=server_ssl_ctx)
-            await site.start()
-
-            port = site._server.sockets[0].getsockname()[1]  # type: ignore[union-attr]
-
-            try:
-                with patch("application_sdk.clients.ssl_utils.SSL_CERT_DIR", tmpdir):
-                    client_ssl_context = get_ssl_context()
-                    assert isinstance(client_ssl_context, ssl.SSLContext)
-
-                    async with httpx.AsyncClient(verify=client_ssl_context) as client:
-                        response = await client.get(
-                            f"https://localhost:{port}/", timeout=10
-                        )
-                        assert response.status_code == 200
-                        assert response.text == "Hello from private HTTPS server!"
-            finally:
-                await runner.cleanup()
-
-    @pytest.mark.asyncio
-    async def test_aiohttp_connects_to_private_https_server(
-        self, trustme_ca, server_ssl_ctx
-    ):  # type: ignore[no-untyped-def]
-        """Test that aiohttp can connect to an HTTPS server using a private CA certificate."""
-        with tempfile.TemporaryDirectory() as tmpdir:
-            ca_cert_path = f"{tmpdir}/ca.pem"
-            trustme_ca.cert_pem.write_to_path(ca_cert_path)  # type: ignore[no-untyped-call]
-
-            async def handle_request(request):  # type: ignore[no-untyped-def]
-                return web.Response(text="Hello from private HTTPS server!")
-
-            app = web.Application()
-            app.router.add_get("/", handle_request)
-
-            runner = web.AppRunner(app)
-            await runner.setup()
-
-            site = web.TCPSite(runner, "127.0.0.1", 0, ssl_context=server_ssl_ctx)
-            await site.start()
-
-            port = site._server.sockets[0].getsockname()[1]  # type: ignore[union-attr]
-
-            try:
-                with patch("application_sdk.clients.ssl_utils.SSL_CERT_DIR", tmpdir):
-                    client_ssl_context = get_ssl_context()
-                    assert isinstance(client_ssl_context, ssl.SSLContext)
-
-                    async with aiohttp.ClientSession() as session:
-                        async with session.get(
-                            f"https://localhost:{port}/",
-                            ssl=client_ssl_context,
-                            timeout=aiohttp.ClientTimeout(total=10),
-                        ) as response:
-                            assert response.status == 200
-                            text = await response.text()
-                            assert text == "Hello from private HTTPS server!"
-            finally:
-                await runner.cleanup()
-
-    @pytest.mark.asyncio
-    async def test_connection_fails_without_custom_ca(self, trustme_ca, server_ssl_ctx):  # type: ignore[no-untyped-def]
-        """Test that connection fails when custom CA is not loaded."""
-
-        async def handle_request(request):  # type: ignore[no-untyped-def]
-            return web.Response(text="Hello!")
-
-        app = web.Application()
-        app.router.add_get("/", handle_request)
-
-        runner = web.AppRunner(app)
-        await runner.setup()
-
-        site = web.TCPSite(runner, "127.0.0.1", 0, ssl_context=server_ssl_ctx)
-        await site.start()
-
-        port = site._server.sockets[0].getsockname()[1]  # type: ignore[union-attr]
-
-        try:
-            with pytest.raises(Exception) as exc_info:
-                async with httpx.AsyncClient(verify=True) as client:
-                    await client.get(f"https://localhost:{port}/", timeout=10)
-
-            error_message = str(exc_info.value).lower()
-            assert "certificate" in error_message or "ssl" in error_message
-        finally:
-            await runner.cleanup()
-
-    @pytest.mark.asyncio
-    async def test_private_and_public_certs_work_simultaneously(
-        self, trustme_ca, server_ssl_ctx
-    ):  # type: ignore[no-untyped-def]
-        """Test that both private server AND public URLs work with the same SSL context."""
-        with tempfile.TemporaryDirectory() as tmpdir:
-            ca_cert_path = f"{tmpdir}/ca.pem"
-            trustme_ca.cert_pem.write_to_path(ca_cert_path)  # type: ignore[no-untyped-call]
-
-            async def handle_request(request):  # type: ignore[no-untyped-def]
-                return web.Response(text="Private server response")
-
-            app = web.Application()
-            app.router.add_get("/", handle_request)
-
-            runner = web.AppRunner(app)
-            await runner.setup()
-
-            site = web.TCPSite(runner, "127.0.0.1", 0, ssl_context=server_ssl_ctx)
-            await site.start()
-
-            port = site._server.sockets[0].getsockname()[1]  # type: ignore[union-attr]
-
-            try:
-                with patch("application_sdk.clients.ssl_utils.SSL_CERT_DIR", tmpdir):
-                    client_ssl_context = get_ssl_context()
-                    assert isinstance(client_ssl_context, ssl.SSLContext)
-
-                    async with httpx.AsyncClient(verify=client_ssl_context) as client:
-                        private_response = await client.get(
-                            f"https://localhost:{port}/", timeout=10
-                        )
-                        assert private_response.status_code == 200
-                        assert private_response.text == "Private server response"
-            finally:
-                await runner.cleanup()
